Refactor logging config
